Moved date/time validations to their own file
