+++ Tewl/InputValidation/Validator.cs
@@ -1,6 +1,4 @@
-﻿using System.Globalization;
-
-namespace Tewl.InputValidation;
+﻿namespace Tewl.InputValidation;
 
 /// <summary>
 /// Contains high-level validation methods. Each validation method returns an object that is the value of the validated
@@ -11,9 +9,6 @@ namespace Tewl.InputValidation;
 /// </summary>
 [ PublicAPI ]
 public class Validator {
-	internal static readonly DateTime SqlSmallDateTimeMinValue = new DateTime( 1900, 1, 1 );
-	internal static readonly DateTime SqlSmallDateTimeMaxValue = new DateTime( 2079, 6, 6 );
-
 	/// <summary>
 	/// Most of our SQL Server decimal columns are specified as (9,2). This is the minimum value that will fit in such a
 	/// column.
@@ -112,207 +107,6 @@ public class Validator {
 		foreach( var message in messages )
 			NoteErrorAndAddMessage( message );
 	}
-
-	/// <summary>
-	/// Returns the validated DateTime type from the given string and validation package.
-	/// It is restricted to the Sql SmallDateTime range of 1/1/1900 up to 6/6/2079.
-	/// Passing an empty string or null will result in ErrorCondition.Empty.
-	/// </summary>
-	public ValidationResult<DateTime> GetSqlSmallDateTime( ValidationErrorHandler? errorHandler, string input ) =>
-		ExecuteValidation<DateTime, string>(
-			errorHandler,
-			input,
-			false,
-			( valueSetter, trimmedInput ) => validateDateTime( valueSetter, trimmedInput, null, SqlSmallDateTimeMinValue, SqlSmallDateTimeMaxValue ) );
-
-	/// <summary>
-	/// Returns the validated DateTime type from the given string and validation package.
-	/// It is restricted to the Sql SmallDateTime range of 1/1/1900 up to 6/6/2079.
-	/// If allowEmpty is true and the given string is empty, null will be returned.
-	/// </summary>
-	public ValidationResult<DateTime?> GetNullableSqlSmallDateTime( ValidationErrorHandler? errorHandler, string input, bool allowEmpty ) =>
-		ExecuteValidation<DateTime?, string>(
-			errorHandler,
-			input,
-			allowEmpty,
-			( valueSetter, trimmedInput ) => validateDateTime(
-				value => valueSetter( value ),
-				trimmedInput,
-				null,
-				SqlSmallDateTimeMinValue,
-				SqlSmallDateTimeMaxValue ) );
-
-	/// <summary>
-	/// Returns the validated DateTime type from the given date part strings and validation package.
-	/// It is restricted to the Sql SmallDateTime range of 1/1/1900 up to 6/6/2079.
-	/// Passing an empty string or null for each date part will result in ErrorCondition.Empty.
-	/// Passing an empty string or null for only some date parts will result in ErrorCondition.Invalid.
-	/// </summary>
-	public ValidationResult<DateTime> GetSqlSmallDateTimeFromParts( ValidationErrorHandler? errorHandler, string month, string day, string year ) =>
-		GetSqlSmallDateTime( errorHandler, makeDateFromParts( month, day, year ) );
-
-	/// <summary>
-	/// Returns the validated DateTime type from the given date part strings and validation package.
-	/// It is restricted to the Sql SmallDateTime range of 1/1/1900 up to 6/6/2079.
-	/// If allowEmpty is true and each date part string is empty, null will be returned.
-	/// Passing an empty string or null for only some date parts will result in ErrorCondition.Invalid.
-	/// </summary>
-	public ValidationResult<DateTime?> GetNullableSqlSmallDateTimeFromParts(
-		ValidationErrorHandler? errorHandler, string month, string day, string year, bool allowEmpty ) =>
-		GetNullableSqlSmallDateTime( errorHandler, makeDateFromParts( month, day, year ), allowEmpty );
-
-	private static string makeDateFromParts( string month, string day, string year ) {
-		var date = month + '/' + day + '/' + year;
-		if( date == "//" )
-			date = "";
-		return date;
-	}
-
-	/// <summary>
-	/// Returns the validated DateTime type from a date string and an exact match pattern.'
-	/// Pattern specifies the date format, such as "MM/dd/yyyy".
-	/// </summary>
-	public ValidationResult<DateTime> GetSqlSmallDateTimeExact( ValidationErrorHandler? errorHandler, string input, string pattern ) =>
-		ExecuteValidation<DateTime, string>(
-			errorHandler,
-			input,
-			false,
-			( valueSetter, trimmedInput ) => validateSqlSmallDateTimeExact( valueSetter, trimmedInput, pattern ) );
-
-	/// <summary>
-	/// Returns the validated DateTime type from a date string and an exact match pattern.'
-	/// Pattern specifies the date format, such as "MM/dd/yyyy".
-	/// </summary>
-	public ValidationResult<DateTime?> GetNullableSqlSmallDateTimeExact( ValidationErrorHandler? errorHandler, string input, string pattern, bool allowEmpty ) =>
-		ExecuteValidation<DateTime?, string>(
-			errorHandler,
-			input,
-			allowEmpty,
-			( valueSetter, trimmedInput ) => validateSqlSmallDateTimeExact( value => valueSetter( value ), trimmedInput, pattern ) );
-
-	private static ValidationError? validateSqlSmallDateTimeExact( Action<DateTime> valueSetter, string trimmedInput, string pattern ) {
-		if( !DateTime.TryParseExact( trimmedInput, pattern, Cultures.EnglishUnitedStates, DateTimeStyles.None, out var date ) )
-			return ValidationError.Invalid();
-		if( validateNativeDateTime( date, false, SqlSmallDateTimeMinValue, SqlSmallDateTimeMaxValue ) is {} error )
-			return error;
-
-		valueSetter( date );
-		return null;
-	}
-
-	private static ValidationError? validateDateTime( Action<DateTime> valueSetter, string trimmedInput, string[]? formats, DateTime min, DateTime max ) {
-		DateTime date;
-		try {
-			date = formats is not null
-				       ? DateTime.ParseExact( trimmedInput, formats, null, DateTimeStyles.None )
-				       : DateTime.Parse( trimmedInput, Cultures.EnglishUnitedStates );
-			if( validateNativeDateTime( date, false, min, max ) is {} error )
-				return error;
-		}
-		catch( FormatException ) {
-			return ValidationError.Invalid();
-		}
-		catch( ArgumentOutOfRangeException ) {
-			// Undocumented exception that there are reports of being thrown
-			return ValidationError.Invalid();
-		}
-		catch( ArgumentNullException ) {
-			return ValidationError.Empty();
-		}
-
-		valueSetter( date );
-		return null;
-	}
-
-	private static ValidationError? validateNativeDateTime( DateTime? date, bool allowEmpty, DateTime minDate, DateTime maxDate ) {
-		if( date is null && !allowEmpty )
-			return ValidationError.Empty();
-		if( date.HasValue ) {
-			var minMaxMessage = " It must be between " + minDate.ToDayMonthYearString( false ) + " and " + maxDate.ToDayMonthYearString( false ) + ".";
-			if( date < minDate )
-				return ValidationError.Custom( ValidationErrorType.TooEarly, "The {0} is too early." + minMaxMessage );
-			if( date >= maxDate )
-				return ValidationError.Custom( ValidationErrorType.TooLate, "The {0} is too late." + minMaxMessage );
-		}
-		return null;
-	}
-
-	/// <summary>
-	/// Validates the date using given allowEmpty, min, and max constraints.
-	/// </summary>
-	public ValidationResult<DateTime?> GetNullableDateTime(
-		ValidationErrorHandler? handler, string input, string[]? formats, bool allowEmpty, DateTime minDate, DateTime maxDate ) =>
-		ExecuteValidation<DateTime?, string>(
-			handler,
-			input,
-			allowEmpty,
-			( valueSetter, trimmedInput ) => validateDateTime( value => valueSetter( value ), trimmedInput, formats, minDate, maxDate ) );
-
-	/// <summary>
-	/// Validates the date using given min and max constraints.
-	/// </summary>
-	public ValidationResult<DateTime> GetDateTime( ValidationErrorHandler? handler, string input, string[]? formats, DateTime minDate, DateTime maxDate ) =>
-		ExecuteValidation<DateTime, string>(
-			handler,
-			input,
-			false,
-			( valueSetter, trimmedInput ) => validateDateTime( valueSetter, trimmedInput, formats, minDate, maxDate ) );
-
-	/// <summary>
-	/// Validates the given time span.
-	/// </summary>
-	public ValidationResult<TimeSpan?> GetNullableTimeSpan( ValidationErrorHandler? handler, TimeSpan? input, bool allowEmpty ) =>
-		ExecuteValidation<TimeSpan?, TimeSpan?>(
-			handler,
-			input,
-			allowEmpty,
-			( valueSetter, trimmedInput ) => {
-				valueSetter( trimmedInput!.Value );
-				return null;
-			} );
-
-	/// <summary>
-	/// Validates the given time span.
-	/// </summary>
-	public ValidationResult<TimeSpan> GetTimeSpan( ValidationErrorHandler? handler, TimeSpan? input ) =>
-		ExecuteValidation<TimeSpan, TimeSpan?>(
-			handler,
-			input,
-			false,
-			( valueSetter, trimmedInput ) => {
-				valueSetter( trimmedInput!.Value );
-				return null;
-			} );
-
-	/// <summary>
-	/// Validates the given time span.
-	/// </summary>
-	public ValidationResult<TimeSpan?> GetNullableTimeOfDayTimeSpan( ValidationErrorHandler? handler, string input, string[]? formats, bool allowEmpty ) =>
-		ExecuteValidation<TimeSpan?, string>(
-			handler,
-			input,
-			allowEmpty,
-			( valueSetter, trimmedInput ) => validateDateTime(
-				value => valueSetter( value.TimeOfDay ),
-				trimmedInput,
-				formats,
-				DateTime.MinValue,
-				DateTime.MaxValue ) );
-
-	/// <summary>
-	/// Validates the given time span.
-	/// </summary>
-	public ValidationResult<TimeSpan> GetTimeOfDayTimeSpan( ValidationErrorHandler? handler, string input, string[]? formats ) =>
-		ExecuteValidation<TimeSpan, string>(
-			handler,
-			input,
-			false,
-			( valueSetter, trimmedInput ) => validateDateTime(
-				value => valueSetter( value.TimeOfDay ),
-				trimmedInput,
-				formats,
-				DateTime.MinValue,
-				DateTime.MaxValue ) );
 
 	/// <summary>
 	/// Executes a validation and returns the result.
